Todos Location
Added method to return user id from auth state synchronously.
Changed path for todos data to include userId instead of passing in the userid and filtering.
Added references to incomplete and complete lists of todos in Todo Service.
Added and refactored interfaces for Todo, CompletedTodo and ObjectData from Firebase.
Updated todo-form to not pass in userId.